+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Fixed
 
 - `ui/VirtualList` not to jump to the top when item size was changed
+ 
 ## [5.4.1] - 2025-12-30
 
 No significant changes.