+++ packages/ui/CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Fixed
 
+- `ui/VirtulList` not to cut off focused item when scroll by pressing key
 - `ui/VirtualList` not to jump to the top when item size was changed
 - `ui/VirtualList.VirtualGridList` to adjust itemSize properly when resizing window